fix night mode function
