+++ features/settings/src/main/java/com/techcafe/todone/settings/di/modules.kt
@@ -1,6 +1,5 @@
 package com.techcafe.todone.settings.di
 
-import com.techcafe.todone.settings.SettingsViewModel
 import org.koin.android.viewmodel.dsl.viewModel
 import org.koin.dsl.module
